refactor: update imports for new client classes location
