+++ cicd/scripts/validate-tag.ts
@@ -1,6 +1,6 @@
 import { ITagModel } from "../core/Models/ITagModel.ts";
-import { RepoClient } from "../core/RepoClient.ts";
-import { TagClient } from "../core/TagClient.ts";
+import { RepoClient } from "../clients/RepoClient.ts";
+import { TagClient } from "../clients/TagClient.ts";
 import { Utils } from "../core/Utils.ts";
 
 const scriptName = Utils.getScriptName();
